testing files: mockResolvedValue params sent as Object
